+++ objects/products.py
@@ -1,0 +1,7 @@
+class Products:
+    def __init__(self, name: str, calories: int, satisfaction_bonus: float, complexity: float):
+        self.name = name
+        self.calories = calories
+        self.satisfaction_bonus = satisfaction_bonus
+        self.manufacturing_complexity = complexity
+        self.complexity = complexity
