+++ settings/constants.py
@@ -1,0 +1,17 @@
+MANUFACTURER_NAMES = ["Manufacturer1", "Manufacturer2"]
+PRODUCT_NAMES = ["cereal", "bread", "milk", "meat", "pie"]
+PRODUCT_CALORIES = [550, 1430, 400, 1820, 2700]
+PRODUCT_COMPLEXITIES = [0.3, 0.5, 0.4, 0.7, 1]
+PRODUCT_BONUSES = [0.5, 1, 1, 1.7, 2.5]
+REQUIRES = [0, 10, 2000, 2]
+PRODUCT_FIRST_PRICE = {"cereal": 0.1, "bread": 0.3, "milk": 0.2, "meat": 0.5, "pie": 1}
+PRODUCTS_COUNT = len(PRODUCT_NAMES)
+SELLERS_COUNT = 15
+BUYERS_COUNT = 80
+MANUFACTURERS_COUNT = 2
+INITIAL_SALARY = 4
+TICKS = 1000
+BUYER_BRAIN_CONSTANT = 10
+BUYER_MEMORY_LEN_CONSTANT = 20
+MANUFACTURER_SALARY_LOW_CONSTANT = 10
+MANUFACTURER_SALARY_UP_CONSTANT = 50
