--- test.py
@@ -1,2 +1,0 @@
-a = [1, 2, 3, 4]
-print(a[:3])
